[runtime] State the prefix crash model; widen page_size to u32
The canonical-prefix shape is a model, not a filesystem guarantee:
device writeback before the creation sync completes may persist bytes
out of order, and such non-prefix tears stay loudly corrupt rather
than healing. Document that explicitly instead of claiming interrupted
writes only leave prefixes, and note in the parse tests that a version
mismatch requires a complete region as well as a valid CRC.

paged::page_size now takes a u32 so the 65536-byte physical page (the
largest whose logical size fits a u16) is expressible.

+++ runtime/src/storage/iouring.rs
@@ -518,7 +518,7 @@ mod tests {
 
         // A logical page size whose physical page is exactly one 4096-byte storage page.
         const PHYSICAL_PAGE_SIZE: u64 = 4096;
-        let logical = crate::buffer::paged::page_size(PHYSICAL_PAGE_SIZE as u16);
+        let logical = crate::buffer::paged::page_size(PHYSICAL_PAGE_SIZE as u32);
         let mut registry = Registry::default();
         let cache = crate::buffer::paged::CacheRef::new(
             test_pool(&mut registry.sub_registry("pool")),
